Moved display functions from run.cpp to display_functions folder
